personare/requirements/test.txt
pytest
tox
coverage
pytest-cov
pytest-flask